+++ DnsCore/Model/Encoding/Data/DnsRecordAddressDataEncoder.cs
@@ -10,7 +10,7 @@ internal sealed class DnsRecordAddressDataEncoder : DnsRecordDataEncoder<IPAddre
 {
     public static readonly DnsRecordAddressDataEncoder Instance = new();
 
-    protected override void EncodeData(ref DnsWriter writer, IPAddress data) => data.TryWriteBytes(writer.ProvideBufferAndAdvance(data.AddressFamily == AddressFamily.InterNetwork ? 4 : 16), out _);
+    protected override void EncodeData(ref DnsWriter writer, IPAddress data) => data.TryWriteBytes(writer.ProvideBufferAndAdvance((ushort)(data.AddressFamily == AddressFamily.InterNetwork ? 4 : 16)), out _);
     protected override IPAddress DecodeData(ref DnsReader reader) => new(reader.ReadToEnd());
     protected override DnsRecord<IPAddress> CreateRecord(DnsName name, IPAddress data, DnsRecordType recordType, DnsClass @class, TimeSpan ttl) => new DnsAddressRecord(name, data, ttl);
 }
